fix(storage): use i64 instead of usize for key expiry

+++ src/storage/mod.rs
@@ -75,7 +75,7 @@ impl Storage {
 		Ok(())
 	}
 
-	async fn expire_key(&self, key: &str, expire_seconds: usize) -> Result<()> {
+	async fn expire_key(&self, key: &str, expire_seconds: i64) -> Result<()> {
 		debug!("Expiring key {key} in {expire_seconds}");
 
 		let mut con = self.client.get_async_connection().await?;
